adds create account txn item.  cleans up other account transactions
