+++ src/features/contacts/index.ts
@@ -1,3 +1,4 @@
 export * from "./components"
 export * from "./store"
 export * from "./types"
+export * from "./hooks"
